Refactor to make project BingoCpp use google-style "bingocpp"
