+++ include/bingocpp/constants.h
