+++ include/bingocpp/equation.h
