+++ include/bingocpp/explicit_regression.h
@@ -22,9 +22,9 @@
 
 #include <Eigen/Core>
 
-#include "BingoCpp/equation.h"
-#include "BingoCpp/fitness_function.h"
-#include "BingoCpp/training_data.h"
+#include "bingocpp/equation.h"
+#include "bingocpp/fitness_function.h"
+#include "bingocpp/training_data.h"
 
 namespace bingo {
