+++ include/bingocpp/implicit_regression.h
@@ -19,11 +19,11 @@
 
 #include <Eigen/Dense>
 
-#include "BingoCpp/equation.h"
-#include "BingoCpp/fitness_function.h"
-#include "BingoCpp/training_data.h"
+#include "bingocpp/equation.h"
+#include "bingocpp/fitness_function.h"
+#include "bingocpp/training_data.h"
 
-#include "BingoCpp/utils.h"
+#include "bingocpp/utils.h"
 
 namespace bingo {
