+++ include/bingocpp/training_data.h
