+++ include/bingocpp/utils.h
