+++ include/bingocpp/version.h
