+++ include/bingocpp/version.h.in
