+++ include/python/py_equation.h
@@ -7,7 +7,7 @@
 #include <pybind11/eigen.h>
 #include <pybind11/stl.h>
 
-#include "BingoCpp/equation.h"
+#include "bingocpp/equation.h"
 
 namespace bingo {
