+++ tests/implicit_regression_tests.cpp
@@ -3,7 +3,7 @@
 #include "gtest/gtest.h"
 #include <Eigen/Dense>
 
-#include <BingoCpp/implicit_regression.h>
+#include <bingocpp/implicit_regression.h>
 
 #include "test_fixtures.h"
