+++ tests/test_fixtures.h
@@ -1,8 +1,8 @@
 #ifndef BINGO_TESTS_TEST_FIXTURES_H_
 #define BINGO_TESTS_TEST_FIXTURES_H_
 
-#include <BingoCpp/agraph.h>
-#include <BingoCpp/equation.h>
+#include <bingocpp/agraph.h>
+#include <bingocpp/equation.h>
 
 #include "testing_utils.h"
